backend/requirements/base.txt
Django==5.2.7
djangorestframework==3.16.1
mysqlclient==2.2.7
celery==5.5.3
redis==6.4.0
qdrant-client==1.15.1
langchain==0.3.27
langchain-qdrant==0.2.1
httpx==0.28.1
python-dotenv==1.1.1
